Extract shared DeleteVersionDialog from the quote rail
Pulls the rail's locally-defined delete-draft-vs-delete-version
dialog into its own component so the upcoming header version menu
can reuse it instead of duplicating the same eligibility branch.
